+++ Package.swift
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "HPLayout",
     platforms: [
-        .iOS(.v11), .tvOS(.v11)
+		.iOS(.v11), .tvOS(.v11), .macOS(.v10_12)
     ],
     products: [
         // Products define the executables and libraries produced by a package, and make them visible to other packages.
