+++ Sources/HPLayout/LayoutAnchor.swift
@@ -1,5 +1,10 @@
-import Foundation
+#if canImport(UIKit)
 import UIKit
+#endif
+
+#if canImport(AppKit)
+import AppKit
+#endif
 
 // MARK: - X Axis
