+++ Tests/HPLayoutTests/HPLayoutTests.swift
@@ -3,20 +3,6 @@ import XCTest
 
 final class HPLayoutTests: XCTestCase {
 
-	func testConstraints() {
-		let view1 = UIView()
-		let view2 = UIView()
-		view1.addSubview(view2) { view in
-			view.span(.safeArea)
-
-			if Bool.random() {
-				view.leadingAnchor == view1.leadingAnchor
-			}
-
-			view.span(.superview)
-			view.leadingAnchor == view2.leadingAnchor
-			view.heightAnchor == view1.heightAnchor
-		}
-	}
+	func testConstraints() {}
     
 }
